requirements.txt
pytorch==1.2.0
torchvision==0.4.0
scipy==1.5.0
scikit-learn==0.23.1
tensorboard==2.2.1
tensorboard-x
future



